match: leftover funcs in dll24

+++ include/sys/camera.h
@@ -112,7 +112,7 @@ void func_80002E94(f32 x, f32 y, f32 z, s32 *ox, s32 *oy, s32 *oz);
 void func_80002F88(s32 x, s32 y, s32 z, f32 *ox, f32 *oy, f32 *oz);
 void func_8000302C(Gfx **gdl);
 void func_80003168(Gfx **gdl, Mtx **rspMtxs, s32 x, s32 y, s32 z, f32 scale);
-void func_800032C4(Gfx **gdl, Mtx **rspMtxs, SRT *param3, f32 param4, s32 param5, MtxF *param6);
+void func_800032C4(Gfx **gdl, Mtx **rspMtxs, SRT *param3, f32 param4, f32 param5, MtxF *param6);
 void func_800034CC(f32 x, f32 y, f32 z);
 void func_8000356C(f32 a, f32 b, f32 c);
 void get_vec3_to_camera_normalized(f32 x, f32 y, f32 z, f32 *ox, f32 *oy, f32 *oz);
